perf report: don't crash on --kallsyms with a non-guest kernel
Without this change, the DSOs for kernel modules aren't marked as loaded
when they're split out from the rest of kallsyms, if --kallsyms is used
with a non-guest kernel.  Thus, the first time a module address is
encountered, we try to re-split the module DSO as if it were a kernel
DSO, which runs into a NULL pointer and crashes.

The previous behavior for DSO_TYPE_GUEST_KERNEL is not changed.